hibiscus/pyproject.toml
# Hibiscus v0.9 — EAZR AI Insurance Intelligence Engine
# Copyright (c) 2026 EAZR Digipayments Pvt Ltd. All rights reserved.

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hibiscus"
version = "0.9.0"
description = "Hibiscus — EAZR AI Insurance Intelligence Engine"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",

    # LangGraph orchestration
    "langgraph>=0.2.50",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",

    # LLM providers via LiteLLM
    "litellm>=1.55.0",
    "openai>=1.56.0",
    "anthropic>=0.40.0",

    # HTTP client
    "httpx>=0.27.0",
    "tenacity>=9.0.0",

    # Redis (session memory)
    "redis>=5.0.0",
    "hiredis>=3.0.0",

    # MongoDB (document memory)
    "motor>=3.6.0",
    "pymongo>=4.9.0",

    # PostgreSQL (user profile, outcomes)
    "asyncpg>=0.30.0",
    "sqlalchemy>=2.0.0",

    # Qdrant (RAG + vector memory)
    "qdrant-client>=1.12.0",

    # Neo4j (Knowledge Graph)
    "neo4j>=5.25.0",

    # Embeddings (local inference — no API key required)
    "fastembed>=0.4.0",
    "tiktoken>=0.8.0",

    # Data validation
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=24.4.0",
    "orjson>=3.10.0",
    "numpy>=2.0.0",
    "python-dateutil>=2.9.0",

    # Observability + web search (optional, gracefully degrade without API keys)
    "langsmith>=0.1.0",
    "tavily-python>=0.3.0",

    # Prometheus metrics (optional — all helpers are no-ops if not installed)
    "prometheus_client>=0.21.0",

    # PDF processing (native extraction pipeline)
    "pdfplumber>=0.10.0",
    "PyPDF2>=3.0.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "pdf2image>=1.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.32.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.hatch.build.targets.wheel]
packages = ["hibiscus"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
